pyproject.toml
[build-system]
requires = ["setuptools >= 77.0.3"]
build-backend = "setuptools.build_meta"

[project]
name = "llama-cpp-py"
version = "0.1.11"
description = "Python wrapper for running the llama.cpp server"
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
license = "MIT"
authors = [
    { name = "sergey21000" }
]

dependencies = [
    "aiohttp==3.13.1",
    "requests==2.32.5",
    "tqdm==4.67.1",
    "openai==2.5.0",
    "python-dotenv==1.1.1",
    "platformdirs==4.5.0"
]

[project.optional-dependencies]
test = [
    "pytest==8.4.2",
    "pytest-asyncio>=1.2.0",
    "colorama==0.4.6",
]

[project.urls]
Homepage = "https://github.com/sergey21000/llama_cpp_py"
Repository = "https://github.com/sergey21000/llama_cpp_py"
